chore: copy over utils to utils file
